requirements.txt
ipykernel
streamlit
langchain
langgraph
python-dotenv
langchain-core
langchain-groq
langchain-openai
langchain-community